requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   django
    #   django-countries
attrs==23.2.0
    # via pytest
autoflake==1.4
    # via -r requirements.in
black==22.3.0
    # via -r requirements.in
bleach[css]==5.0.1
    # via django-bleach
boto3==1.34.102
    # via django-storages
botocore==1.34.102
    # via
    #   boto3
    #   s3transfer
build==1.2.1
    # via pip-tools
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via
    #   requests
    #   sentry-sdk
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   django-click
    #   pip-tools
coverage==7.5.1
    # via pytest-cov
cssselect==1.2.0
    # via pyquery
distlib==0.3.8
    # via virtualenv
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.25
    # via
    #   -r requirements.in
    #   django-admin-sortable2
    #   django-bleach
    #   django-debug-toolbar
    #   django-extensions
    #   django-formtools
    #   django-leaflet
    #   django-recaptcha
    #   django-sendgrid-v5
    #   django-storages
    #   django-tinymce
    #   django-unused-media
    #   easy-thumbnails
django-admin-sortable2==1.0.4
    # via -r requirements.in
django-autoslug==1.9.4
    # via -r requirements.in
django-bleach==3.1.0
    # via -r requirements.in
django-click==2.3.0
    # via -r requirements.in
django-countries==7.6.1
    # via -r requirements.in
django-date-extensions==3.1.2
    # via -r requirements.in
django-debug-toolbar==4.3.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-formtools==2.3.0
    # via -r requirements.in
django-gulp-rev==0.2
    # via -r requirements.in
django-leaflet==0.24.0
    # via -r requirements.in
django-markdown-deux==1.0.5
    # via -r requirements.in
django-recaptcha==4.0.0
    # via -r requirements.in
django-sendgrid-v5==0.8.1
    # via -r requirements.in
django-storages[boto3]==1.7.1
    # via -r requirements.in
django-tinymce==4.0.0
    # via -r requirements.in
django-unused-media==0.1.13
    # via -r requirements.in
dnspython==2.6.1
    # via -r requirements.in
easy-thumbnails==2.8.5
    # via -r requirements.in
easydict==1.10
    # via -r requirements.in
faker==25.1.0
    # via -r requirements.in
filelock==3.14.0
    # via
    #   urlextract
    #   virtualenv
flake8==4.0.1
    # via -r requirements.in
freezegun==1.5.0
    # via -r requirements.in
future==1.0.0
    # via django-sendgrid-v5
google-api-core==2.19.0
    # via google-api-python-client
google-api-python-client==2.37.0
    # via -r requirements.in
google-auth==2.29.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.2.0
    # via google-api-python-client
googleapis-common-protos==1.63.0
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
icalendar==4.0.3
    # via -r requirements.in
identify==2.5.36
    # via pre-commit
idna==3.7
    # via
    #   requests
    #   urlextract
    #   yarl
iniconfig==2.0.0
    # via pytest
isort==5.8.0
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==5.2.1
    # via pyquery
markdown2==2.4.13
    # via django-markdown-deux
mccabe==0.6.1
    # via flake8
multidict==6.0.5
    # via yarl
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pre-commit
oauth2client==4.1.3
    # via -r requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.0
    # via
    #   build
    #   pytest
pathspec==0.12.1
    # via black
pillow==10.2.0
    # via
    #   -r requirements.in
    #   easy-thumbnails
pip-tools==6.12.1
    # via -r requirements.in
platformdirs==4.2.1
    # via
    #   black
    #   urlextract
    #   virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==2.21.0
    # via -r requirements.in
proto-plus==1.23.0
    # via google-api-core
protobuf==4.25.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg2-binary==2.8.3
    # via -r requirements.in
py==1.11.0
    # via pytest
py-trello==0.13.0
    # via -r requirements.in
pyasn1==0.6.0
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via
    #   google-auth
    #   oauth2client
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via
    #   autoflake
    #   flake8
pyparsing==3.1.2
    # via httplib2
pyproject-hooks==1.1.0
    # via build
pyquery==1.4.0
    # via -r requirements.in
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-dotenv
    #   pytest-env
    #   pytest-mock
pytest-cov==2.12.1
    # via -r requirements.in
pytest-django==4.4.0
    # via -r requirements.in
pytest-dotenv==0.5.2
    # via -r requirements.in
pytest-env==0.6.2
    # via -r requirements.in
pytest-mock==3.10.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   faker
    #   freezegun
    #   icalendar
    #   py-trello
python-dotenv==0.21.0
    # via
    #   -r requirements.in
    #   pytest-dotenv
python-http-client==3.3.7
    # via
    #   django-sendgrid-v5
    #   sendgrid
pytz==2024.1
    # via
    #   django
    #   icalendar
    #   py-trello
pyyaml==6.0.1
    # via
    #   pre-commit
    #   vcrpy
requests==2.31.0
    # via
    #   google-api-core
    #   py-trello
    #   requests-oauthlib
    #   stripe
requests-oauthlib==2.0.0
    # via py-trello
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
s3transfer==0.10.1
    # via boto3
sendgrid==6.11.0
    # via django-sendgrid-v5
sentry-sdk==2.1.1
    # via -r requirements.in
six==1.16.0
    # via
    #   bleach
    #   django-unused-media
    #   oauth2client
    #   python-dateutil
slack-sdk==3.19.5
    # via -r requirements.in
sqlparse==0.5.0
    # via
    #   django
    #   django-debug-toolbar
starkbank-ecdsa==2.2.0
    # via sendgrid
stripe==9.6.0
    # via -r requirements.in
tinycss2==1.1.1
    # via bleach
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
tomli==2.0.1
    # via
    #   black
    #   build
typing-extensions==4.11.0
    # via
    #   asgiref
    #   django-countries
    #   stripe
uritemplate==4.1.1
    # via google-api-python-client
uritools==4.0.2
    # via urlextract
urlextract==1.9.0
    # via -r requirements.in
urllib3==2.2.1
    # via
    #   botocore
    #   requests
    #   sentry-sdk
vcrpy==6.0.1
    # via -r requirements.in
virtualenv==20.26.1
    # via pre-commit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
wheel==0.43.0
    # via pip-tools
wrapt==1.16.0
    # via vcrpy
yarl==1.9.4
    # via vcrpy

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
